+++ test/test.js
@@ -168,12 +168,12 @@ describe('RNPickerSelect', () => {
         expect(wrapper.state().showPicker).toEqual(false);
     });
 
-    describe('when provided a displayValue item property', () => {
+    it('should show the value "RED" in the TextInput instead of the label "Red" when the inputLabel is set', () => {
         const items = [
             {
                 label: 'Red',
                 value: 'red',
-                displayValue: true,
+                inputLabel: 'RED',
             },
             {
                 label: 'Orange',
@@ -181,37 +181,13 @@ describe('RNPickerSelect', () => {
             },
         ];
 
-        it('should show the value (`red`) in the TextInput instead of the label (`Red`) when the displayValue is set to true', () => {
-            const wrapper = shallow(
-                <RNPickerSelect
-                    items={items}
-                    placeholder={{}}
-                    onValueChange={noop}
-                    useValueForTextInput
-                    value="red"
-                />
-            );
+        const wrapper = shallow(
+            <RNPickerSelect items={items} placeholder={{}} onValueChange={noop} value="red" />
+        );
 
-            const textInput = wrapper.find('[testID="text_input"]');
+        const textInput = wrapper.find('[testID="text_input"]');
 
-            expect(textInput.props().value).toEqual('red');
-        });
-
-        it('should show the label (`Orange`) in the TextInput instead of the value (`orange`) when the displayValue is set to false or it is not set', () => {
-            const wrapper = shallow(
-                <RNPickerSelect
-                    items={items}
-                    placeholder={{}}
-                    onValueChange={noop}
-                    useValueForTextInput
-                    value="orange"
-                />
-            );
-
-            const textInput = wrapper.find('[testID="text_input"]');
-
-            expect(textInput.props().value).toEqual('Orange');
-        });
+        expect(textInput.props().value).toEqual('RED');
     });
 
     it('should update the selected value when the `value` prop updates and call the onValueChange cb', () => {
@@ -326,6 +302,17 @@ describe('RNPickerSelect', () => {
 
         const component = wrapper.find('[testID="android_picker_headless"]');
         expect(component).toHaveLength(1);
+    });
+
+    it('should set the selected value to state (Web)', () => {
+        Platform.OS = 'web';
+        const wrapper = shallow(<RNPickerSelect items={selectItems} onValueChange={noop} />);
+
+        wrapper
+            .find('[testID="web_picker"]')
+            .props()
+            .onValueChange('orange', 2);
+        expect(wrapper.state().selectedItem.value).toEqual('orange');
     });
 
     it('should call the onDonePress callback when set (iOS)', () => {